+++ src/app/_components/myfamily/Calendar.module.scss
@@ -1,0 +1,4 @@
+.main-calander-container {
+  margin-top: 2.6rem;
+  padding: 0 2.4rem;
+}
