+++ src/app/_components/myfamily/Checklist.module.scss
@@ -1,0 +1,38 @@
+.checklist {
+  @include column-flexbox(center, start, 1rem);
+  padding: 0 2.4rem;
+  width: 100%;
+
+  &-title {
+    @include text-style(18, $black, bolder);
+    @include flexbox(center, center, 1rem);
+    line-height: 3.3rem;
+  }
+
+  &-main {
+    @include text-style(16, $black, light);
+    @include column-flexbox(center, start, 0);
+    width: 100%;
+    padding: 0.3rem 0.9rem 0.3rem 0.9rem;
+    border-radius: 0.6rem;
+    background: $white;
+    box-shadow: 0 0.4rem 0.4rem 0 rgba(0, 0, 0, 0.25);
+    line-height: 3.3rem;
+
+    > .checklist-main-section:not(:last-child) {
+      border-bottom: 0.1rem solid $gray20;
+    }
+
+    &-section {
+      @include flexbox(between, center, 0);
+      width: 100%;
+      padding-left: 1.4rem;
+    }
+
+    &-menu {
+      @include flexbox(center, center);
+      height: 3.3rem;
+      padding-left: 1.4rem;
+    }
+  }
+}
