+++ src/app/_components/myfamily/MainCalendar.scss
@@ -1,11 +1,3 @@
-/* 일요일 색상 */
-.fc-day-sun a {
-}
-
-/* 토요일 색상 */
-.fc-day-sat a {
-}
-
 .fc-media-screen {
   border-radius: 0.6rem;
   border: 0.1rem solid $primary-10;
@@ -156,3 +148,4 @@
 .fc .fc-daygrid-day {
   height: 3.4rem;
 }
+// }
