+++ src/app/_components/myfamily/TravelRecord.module.scss
@@ -1,0 +1,65 @@
+.travelRecord {
+  @include column-flexbox(center, start, 1rem);
+  width: 100%;
+  margin-top: 0.3rem;
+  padding-left: 2.4rem;
+
+  &-title {
+    @include text-style(18, $black, bolder);
+    @include flexbox(center, center, 1rem);
+    line-height: 3.3rem;
+  }
+
+  .scroll-container {
+    width: 100%;
+    display: flex;
+    flex-wrap: nowrap;
+    overflow-x: scroll;
+    -webkit-overflow-scrolling: touch;
+    scroll-behavior: smooth;
+
+    &::-webkit-scrollbar {
+      height: 1.2rem;
+    }
+
+    &::-webkit-scrollbar-track {
+      background: transparent;
+    }
+
+    &::-webkit-scrollbar-thumb {
+      background-color: $gray10;
+      border-radius: 0.8rem;
+      border: 0.2rem solid transparent;
+      background-clip: padding-box;
+    }
+
+    &::-webkit-scrollbar-thumb:hover {
+      background-color: $gray;
+      transition: $base-transition;
+    }
+  }
+
+  &-main {
+    @include flexbox(start, center, 1.8rem);
+    margin: 0.4rem;
+    width: 100%;
+    white-space: nowrap;
+
+    &-section {
+      @include column-flexbox(between, start, 1.8rem);
+      @include text-style(12, $black, light);
+      width: 14.8rem;
+      height: 12.4rem;
+      padding: 0.4rem 0 1.7rem 2rem;
+      border-radius: 0.6rem;
+      background: $white;
+      box-shadow: 0 0.4rem 0.4rem 0 rgba(0, 0, 0, 0.25);
+      flex-shrink: 0;
+
+      &-title {
+        @include text-style(16, $black, bolder);
+        line-height: 3.3rem;
+      }
+    }
+  }
+}
