+++ src/app/_components/myfamily/Upcoming.module.scss
@@ -1,0 +1,22 @@
+.upcoming {
+  @include column-flexbox(center, start, 0.5rem);
+  width: 100%;
+  margin-top: 0.6rem;
+  padding: 0 2.4rem;
+
+  &-title {
+    @include text-style(18, $black, bolder);
+    @include flexbox(center, center, 1rem);
+    line-height: 3.3rem;
+  }
+
+  &-main {
+    @include text-style(16, $black, light);
+    @include flexbox(between, between, 1rem);
+    width: 100%;
+    padding: 1.5rem 3.6rem 1.4rem 2.3rem;
+    border-radius: 0.6rem;
+    background: $white;
+    box-shadow: 0 0.4rem 0.4rem 0 rgba(0, 0, 0, 0.25);
+  }
+}
