+++ src/app/myfamily/MyFamily.module.scss
@@ -1,153 +1,4 @@
 .container {
   @include column-flexbox(start, start, 1.2rem);
-
-  // padding: 8rem 0 0 2.4rem;
   width: 100%;
-}
-.calander-container {
-  margin-top: 2.6rem;
-  padding: 0 2.4rem;
-}
-
-.upcoming {
-  @include column-flexbox(center, start, 0.5rem);
-
-  width: 100%;
-  margin-top: 0.6rem;
-  padding: 0 2.4rem;
-
-  &-title {
-    @include text-style(18, $black, bolder);
-    @include flexbox(center, center, 1rem);
-
-    line-height: 3.3rem;
-  }
-
-  &-main {
-    @include text-style(16, $black, light);
-    @include flexbox(between, between, 1rem);
-
-    width: 100%;
-    padding: 1.5rem 3.6rem 1.4rem 2.3rem;
-    border-radius: 0.6rem;
-    background: $white;
-    box-shadow: 0 0.4rem 0.4rem 0 rgba(0, 0, 0, 0.25);
-  }
-}
-
-.checklist {
-  @include column-flexbox(center, start, 1rem);
-  padding: 0 2.4rem;
-
-  width: 100%;
-
-  &-title {
-    @include text-style(18, $black, bolder);
-    @include flexbox(center, center, 1rem);
-
-    line-height: 3.3rem;
-  }
-
-  &-main {
-    @include text-style(16, $black, light);
-    @include column-flexbox(center, start, 0);
-
-    width: 100%;
-    padding: 0.3rem 0.9rem 0.3rem 0.9rem;
-    border-radius: 0.6rem;
-    background: $white;
-    box-shadow: 0 0.4rem 0.4rem 0 rgba(0, 0, 0, 0.25);
-    line-height: 3.3rem;
-
-    > .checklist-main-section:not(:last-child) {
-      border-bottom: 0.1rem solid $gray20;
-    }
-
-    &-section {
-      @include flexbox(between, center, 0);
-
-      width: 100%;
-      padding-left: 1.4rem;
-    }
-
-    &-menu {
-      @include flexbox(center, center);
-
-      height: 3.3rem;
-      padding-left: 1.4rem;
-    }
-  }
-}
-
-.travelRecord {
-  @include column-flexbox(center, start, 1rem);
-
-  width: 100%;
-  margin-top: 0.3rem;
-  padding-left: 2.4rem;
-
-  &-title {
-    @include text-style(18, $black, bolder);
-    @include flexbox(center, center, 1rem);
-
-    line-height: 3.3rem;
-  }
-
-  .scroll-container {
-    width: 100%;
-    display: flex;
-    flex-wrap: nowrap;
-    overflow-x: scroll;
-
-    -webkit-overflow-scrolling: touch;
-    scroll-behavior: smooth;
-
-    /* 스크롤바 스타일 커스터마이징 */
-    &::-webkit-scrollbar {
-      height: 1.2rem; /* 스크롤바 높이 설정 */
-    }
-
-    &::-webkit-scrollbar-track {
-      background: transparent; /* 스크롤바 배경 투명 처리 */
-    }
-
-    &::-webkit-scrollbar-thumb {
-      background-color: $gray10; /* 스크롤바 색상 */
-      border-radius: 0.8rem; /* 스크롤바에 border-radius 추가 */
-      border: 0.2rem solid transparent; /* 스크롤바에 투명한 테두리 추가 */
-      background-clip: padding-box; /* 테두리와 thumb 사이에 공간 생성 */
-    }
-
-    &::-webkit-scrollbar-thumb:hover {
-      background-color: $gray; /* 호버 시 스크롤바 색상 변경 */
-      transition: $base-transition;
-    }
-  }
-
-  &-main {
-    @include flexbox(start, center, 1.8rem);
-
-    margin: 0.4rem;
-    width: 100%;
-    white-space: nowrap;
-
-    &-section {
-      @include column-flexbox(between, start, 1.8rem);
-      @include text-style(12, $black, light);
-
-      width: 14.8rem;
-      height: 12.4rem;
-      padding: 0.4rem 0 1.7rem 2rem;
-      border-radius: 0.6rem;
-      background: $white;
-      box-shadow: 0 0.4rem 0.4rem 0 rgba(0, 0, 0, 0.25);
-      flex-shrink: 0;
-
-      &-title {
-        @include text-style(16, $black, bolder);
-
-        line-height: 3.3rem;
-      }
-    }
-  }
 }
